Docs for wrapping wc stories + lit-html dev dep
Issue:

## What I did

### Wrapping docs

Didn't see this in any official docs, but thought an example for wrapping stories in Web Component-land would be beneficial. I think I had to dig through some Github issues to find an implementation. 

Not sure if the FAQ was the proper place for it, but thought it'd be an ok place for now? 👌

### lit-html dependency

Not sure I used the Storybook init or the OpenWC version, but I had an issue where lit-html wasn't included as a dev-dependency and it crashed the project on a fresh install. It should be made clear somewhere (maybe later in [more details "manual" instructions](https://github.com/storybookjs/storybook/issues/9693)) that it's needed. 

The FAQ seemed like a proper place, since the "Getting Started" section uses the CLI script, which ideally should install it? Once there's a manual section, it can be moved there instead.

## How to test

- Is this testable with Jest or Chromatic screenshots?: No
- Does this need a new example in the kitchen sink apps?: No
- Does this need an update to the documentation?: I don't think so.

<!--

Everybody: Please submit all PRs to the `next` branch unless they are specific to current release. Storybook maintainers cherry-pick bug and documentation fixes into the `master` branch as part of the release process, so you shouldn't need to worry about this.

Maintainers: Please tag your pull request with at least one of the following:
`["cleanup", "BREAKING CHANGE", "feature request", "bug", "documentation", "maintenance", "dependencies", "other"]`

-->